requirements.txt
vector
awkward
numpy
particle
fastjet
uproot
hydra-core
numba
law
luigi